Refactor NFTRewardSection
